### 20250720, v0.9.4 Build 672 * **add LLVM/CLANG-cl-family support — the library source code itself is now buildable with LLVM/CLANG-cl:**     * support VS2022 v17.14 with **LLVM (clang-cl)** tool chain     * support **Intel C++ Compiler 2024** tool chain<br>       **NOTE:** Only [**Intel oneAPI Base Toolkir ver 2024.2.1**](https://www.intel.com/content/www/us/en/developer/tools/oneapi/base-toolkit-download.html)       provides the 32Bit compiler.<br>In versions **2025.x.y** the 32Bit compiler isn't shipped anymore. * Windows SDK version **10.0.26100.0** * support Microsoft tool chain:     * **v143(VS2022)**     * **v142(VS2019)**     * **v141(VS2017)**

+++ toroCLibrary/LibCore/_cdeMemStrxCpy.c
@@ -25,7 +25,7 @@ void* _cdeMemStrxCpy(IN int preset, OUT void* pszDest, IN const void* pszSource,
 {
     unsigned n;
     signed xcrement = PRESET_TDN ? -1 : 1;//increment vs decrement
-    unsigned short 	c;
+    unsigned short  c;
     unsigned char* pDst8 = pszDest;
     unsigned short* pDst16 = pszDest;
     const unsigned char* pSrc8 = pszSource;
@@ -49,7 +49,7 @@ void* _cdeMemStrxCpy(IN int preset, OUT void* pszDest, IN const void* pszSource,
     //
     // strncpy support -- that fills remaining buffer until count
     //
-    if ((-1 != count) && PRESET_BOZ) {
+    if (((size_t) -1 != count) && PRESET_BOZ) {
         for ( /*n = 0*/; n < count; n++, pDst16 = &pDst16[xcrement], pDst8 = &pDst8[xcrement]) {
 
             if (PRESET_WID) {
