+++ toroCLibrary/Library/stdio_h/FputcCDEABI.c
@@ -56,7 +56,7 @@ Returns
 
 **/
 static int fputcCDEABI(int b, FILE* stream) {
-    unsigned int nRet = (size_t)EOF;
+    unsigned nRet = (unsigned)EOF;
     unsigned char c = (unsigned char)b;
 
     do {
