+++ toroCLibrary/Library/wchar_h/Ungetwc.c
@@ -92,5 +92,5 @@ wint_t ungetwc(wint_t c, FILE * stream)
             } while (0);
 
     }
-    return nRet == EOF ? WEOF : (wint_t)nRet;
+    return nRet == (unsigned)EOF ? WEOF : (wint_t)nRet;
 }
